+++ CHANGELOG.md
@@ -14,6 +14,9 @@
 - `zoom` values now match standard 256px tile maps. Each zoom level's value increases by 1.
   **Breaking**: update all zoom level dependent styles by adding 1 to each zoom level.
 - Added `Point` class to encapsulate x, y coords.
+  API methods will accept either `[x, y]` or `new llmr.Point(x, y)`.
+- Changed `Map` `panBy` signature to accept `point` instead of `x, y`.
+- Changed `Map` `fitBounds` and `zoomPanTo` to accept `offset` (Point) instead of `offsetX, offsetY`.
 - Raster buckets now need to specify `type: 'raster'` in the stylesheet.
 
 ### 0.0.7